// frontend/frontend-ng/src/app/services/models/internship.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface Internship {
  company?: string;
  endDate?: string;
  id?: number;
  open?: boolean;
  paid?: boolean;
  startDate?: string;
  technology?: 'JAVA' | 'PYTHON' | 'CPP' | 'CS' | 'ANGULAR' | 'REACT' | 'VUEJS';
}